+++ modules/requisition/src/main/java/org/openlmis/rnr/domain/RequisitionGroupMember.java
@@ -1,6 +1,5 @@
 package org.openlmis.rnr.domain;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -9,13 +8,13 @@ import java.util.Date;
 @Data
 @NoArgsConstructor
 public class RequisitionGroupMember {
-    private Integer id;
-    private Integer requisitionGroupId;
-    private Integer facilityId;
+    private Long id;
+    private Long requisitionGroupId;
+    private Long facilityId;
     private String modifiedBy;
     private Date modifiedDate;
 
-    public RequisitionGroupMember(Integer facilityId, Integer requisitionGroupId) {
+    public RequisitionGroupMember(Long facilityId, Long requisitionGroupId) {
         this.facilityId = facilityId;
         this.requisitionGroupId = requisitionGroupId;
     }
